+++ movai_core_shared/__init__.py
@@ -105,6 +105,13 @@ from movai_core_shared.envvars import (
     ENVIRON_ROS2,
     ENVIRON_GDNODE,
     ENVIRON_GDNODE_INJECT,
+    TELEMETRY_ENABLE,
+    MOVAI_TELEMETRY_SOCKET,
+    PLATFORM_METRICS_SOCKET_PATH,
+    MIMIR_HOST,
+    MIMIR_PORT,
+    MIMIR_URL,
+    LOKI_URL,
 )
 
 from movai_core_shared.exceptions import (
@@ -234,4 +241,11 @@ __all__ = [
     "RECOVERY_STATE_KEY",
     "RECOVERY_TIMEOUT_IN_SECS",
     "RECOVERY_RESPONSE_KEY",
+    "TELEMETRY_ENABLE",
+    "MOVAI_TELEMETRY_SOCKET",
+    "PLATFORM_METRICS_SOCKET_PATH",
+    "MIMIR_HOST",
+    "MIMIR_PORT",
+    "MIMIR_URL",
+    "LOKI_URL",
 ]